refactor(web): extract resource rows + share formatBytes
ResourcesSection inlined a pending-resource row and a synced-resource row;
formatBytes was defined twice (here and MessageResourceSection). Extract
PendingResourceRow and SyncedResourceRow, and move formatBytes to lib/. Typing
the synced row consolidates 8 untyped-live-query errors down to 1. ResourcesSection
drops 259->149 lines. Behavior-preserving.
